Rename rdb [ci skip]
